PC - array indexing must be integer

+++ src/Components/missions/PACE/nobm.py
@@ -251,15 +251,15 @@ class NOBM(object):
                         lon[i] = lon[i] + 360.0
 
                         nlat,nlon = datach.shape
-                        lon1 = self.lon[0:nlon*0.5]
-                        lon2 = self.lon[nlon*0.5:]
+                        lon1 = self.lon[0:int(nlon*0.5)]
+                        lon2 = self.lon[int(nlon*0.5):]
                         datalon = np.append(lon2,lon1+360.)
 
                     else:
                         #center at 0. Lon goes from -180 to 180
                         nlat,nlon = datach.shape
-                        data1 = datach[:,0:nlon*0.5]
-                        data2 = datach[:,nlon*0.5:]
+                        data1 = datach[:,0:int(nlon*0.5)]
+                        data2 = datach[:,int(nlon*0.5):]
                         datach  = np.ma.hstack((data2, data1))
                         datalon = self.lon
